requirements.txt
requests>=2.28.1
python-whois>=0.8.0
colorama>=0.4.6